Fix MetricsCollector, resume, aux loss scaling, dtype mismatches
1. MetricsCollector: rewrite for v4 structure. Remove v3 references
   (block.layers, model.wm, model.W_in, gate_ab). PM stats now from
   block.pm directly. Grad norms track fan_out/fan_in/columns/pm/em
   per block.

2. Resume state wipe: set trainer._states_initialized = True after
   loading runtime state, preventing train_chunk() from reinitializing
   and overwriting loaded PM/EM.

3. Aux loss scaling: aux_loss (mean-reduced from F.mse_loss) was added
   to ce_loss (sum-reduced) then divided by valid_count, making PCM
   loss ~1/(BS*N) too small. Now scale aux by seg_valid before adding.

4. PMNeuromodulator disabled path: hardcoded slot_logits width 8,
   now uses config.r via self.n_slots.

5. Dtype mismatches on CUDA: state tensors are bf16 (runtime_state_dtype)
   but model params are fp32. Fixed einsum dtype mismatches in PM.read,
   EM.read, EM.write, and block eligibility routing by casting to state
   dtype before einsum, casting back for learned parameter ops.

+++ src/model/block.py
@@ -51,8 +51,9 @@ class ColumnBlock(nn.Module):
             # Route eligibility to PM slots
             k_norm = unit_normalize(k_cand)
             # pm_K: [BS, r, D_mem], k_norm: [BS, N, C, D_mem]
+            # Cast to state dtype for einsum compatibility (state may be bf16)
             route_scores = torch.einsum(
-                "bncd, brd -> bncr", k_norm, self.pm.pm_K
+                "bncd, brd -> bncr", k_norm.to(self.pm.pm_K.dtype), self.pm.pm_K
             )  # [BS, N, C, r]
             route_w = torch.softmax(
                 route_scores / self.config.tau_route_pm, dim=-1
